+++ src/server.js
@@ -10,12 +10,11 @@ const authApi = require("./routes/auth");
 const uploadApi = require("./routes/upload");
 const clubApi = require("./routes/club");
 const notificationApi = require("./routes/notification");
-const noticeApi = require("./routes/notice");
+
 const boardApi = require("./routes/board");
-const postApi = require("./routes/post");
-const commentApi = require("./routes/comment");
-const replyApi = require("./routes/reply");
-const promotionApi = require("./routes/promotion");
+const general = require("./routes/general");
+const promotion = require("./routes/promotion");
+const notice = require("./routes/notice/notice");
 
 const errorHandling = require("./middleware/errorHandling");
 
@@ -32,12 +31,11 @@ app.use("/auth", authApi);
 app.use("/upload", uploadApi);
 app.use("/club", clubApi);
 app.use("/notification", notificationApi);
-app.use("/notice", noticeApi);
 app.use("/board", boardApi);
-app.use("/post", postApi);
-app.use("/comment", commentApi);
-app.use("/reply", replyApi);
-app.use("/promotion", promotionApi);
+
+app.use("/general", general);
+app.use("/notice", notice);
+app.use("/promotion", promotion);
 
 // error handling muddleware
 app.use(errorHandling());
